backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-docx
requests
google-genai
python-multipart
